+++ NEWS.md
@@ -4,6 +4,10 @@ We are preparing to release the package on CRAN and are making relevant changes 
 
 # What's new in SITS version 0.15.1
 
+### New features in SITS version 0.15.1-1
+* Fix bug in `sits_regularize` that generated images with artifacts
+* Fix wrong *bbox* in `sits_cube` from STAC AWS Sentinel-2
+
 ### New features in SITS version 0.15.1
 * Update README.Rmd
 * Support `sits_timeline()` to sits model objects